pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pathlib>=1.0.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "werkzeug>=3.1.3",
]
